restructured Pipeline class as class factory
