+++ sam_ml/models/__init__.py
@@ -1,0 +1,5 @@
+from .main_pipeline import Pipeline
+
+__all__ = {
+    "pipeline class": "Pipeline",
+}
